defi_agent_eval/requirements.txt
# Core dependencies
web3==6.11.3
eth-account==0.10.0
eth-typing==3.5.2
eth-utils==2.3.1

# Blockchain interaction
py-solc-x==2.0.2

# OpenAI for LLM agent
openai==1.3.0
anthropic==0.7.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0

# Development
pytest==7.4.3
black==23.11.0
